+++ src/components/manage/Sidebar/ObjectBrowserBody.jsx
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { defineMessages, FormattedMessage, injectIntl } from 'react-intl';
-import { Input, Segment } from 'semantic-ui-react';
+import { Input, Segment, Breadcrumb } from 'semantic-ui-react';
+
 import { join } from 'lodash';
 
 // These absolute imports (without using the corresponding centralized index.js) are required
@@ -19,6 +20,7 @@ import folderSVG from '@plone/volto/icons/folder.svg';
 import clearSVG from '@plone/volto/icons/clear.svg';
 import searchSVG from '@plone/volto/icons/zoom.svg';
 import linkSVG from '@plone/volto/icons/link.svg';
+import homeSVG from '@plone/volto/icons/home.svg';
 
 import ObjectBrowserNav from '@plone/volto/components/manage/Sidebar/ObjectBrowserNav';
 
@@ -393,7 +395,38 @@ class ObjectBrowserBody extends Component {
             <Icon name={clearSVG} size="24px" />
           </button>
         </header>
-        <Segment secondary>{this.state.currentFolder}</Segment>
+        <Segment secondary className="breadcrumbs" vertical>
+          <Breadcrumb>
+            {this.state.currentFolder !== '/' ? (
+              this.state.currentFolder.split('/').map((item, index, items) => {
+                return (
+                  <React.Fragment key={`divider-${item}-${index}`}>
+                    {index === 0 ? (
+                      <Breadcrumb.Section onClick={() => this.navigateTo('/')}>
+                        <Icon name={homeSVG} size="18px" />
+                      </Breadcrumb.Section>
+                    ) : (
+                      <>
+                        <Breadcrumb.Divider key={`divider-${item.url}`} />
+                        <Breadcrumb.Section
+                          onClick={() =>
+                            this.navigateTo(items.slice(0, index + 1).join('/'))
+                          }
+                        >
+                          {item}
+                        </Breadcrumb.Section>
+                      </>
+                    )}
+                  </React.Fragment>
+                );
+              })
+            ) : (
+              <Breadcrumb.Section onClick={() => this.navigateTo('/')}>
+                <Icon name={homeSVG} size="18px" />
+              </Breadcrumb.Section>
+            )}
+          </Breadcrumb>
+        </Segment>
         {this.props.mode === 'multiple' && (
           <Segment className="infos">
             {this.props.intl.formatMessage(messages.SelectedItems)}:{' '}