+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Feature
 - Allowing user to paste url in search box in objectBrowser @iFlameing
+- Allowing user to click on the breadcrumbs of objectBrowser @iFlameing
 
 ### Bugfix